fix: changed type for balance

+++ FinancialTracker.Frontend/Models/Account.cs
@@ -5,7 +5,7 @@
 		public Guid AccountId { get; set; }
 		public Guid UserId { get; set; }
 		public string Name { get; set; }
-		public double CurrentBalance { get; set; }
+		public decimal CurrentBalance { get; set; }
 		public int Currency { get; set; }
 		public string CreatedAt { get; set; }
 		public string? UpdatedAt { get; set; }
